+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.0.7
   * Only update line dash value when necessary
   * Upgraded to `polymer 1.0.0-rc.14`, `polymer-elements 1.0.0-rc.7` and `reflectable 0.5.1`
+  * __+1__: Clear intersection indices for every child; improve intersection efficiency
 
 ### 0.0.6
   * Added a default line dash value